refactor: remove aliasedExpression from GroupBy
